feat(app-preview)!: add exclude glob support (#156)
