+++ apps/preview/index.d.ts
@@ -1,0 +1,1 @@
+export const getDeps: () => string[];
