+++ apps/preview/package.json
@@ -19,10 +19,12 @@
   },
   "files": [
     "app",
+    "index.d.ts",
     "index.js",
     "README.md"
   ],
   "peerDependencies": {
+    "micromatch": "^4.0.5",
     "react": "^18.2.0"
   },
   "dependencies": {
@@ -44,5 +46,8 @@
     "titleize": "^4.0.0",
     "vite": "^4.4.9",
     "vite-plugin-node-polyfills": "^0.16.0"
-  }
+  },
+  "optimizeDeps": [
+    "micromatch"
+  ]
 }
